Rectangle rendering for even y coordinates

+++ src/types.c
@@ -1,7 +1,7 @@
 #include "types.h"
 
 
-Vector2 vec2(i32 x, i32 y) {
+Vector2 vec2(i16 x, i16 y) {
 
     Vector2 v;
     v.x = x;
